+++ src/controllers/ethereum/_tests_/contract.spec.ts
@@ -2,6 +2,7 @@ import 'jest';
 import {__consumerInstance__} from '../../../domain/consumers/__mocks__/consumer';
 import {findOne} from '../../../domain/ethereum';
 import {IEthContractEventBody} from '../../../models/ethereum';
+import {CONSUMER_EVENT_KINDS, CURRENCY} from '../../../models/models';
 import {onError} from '../../../utils/error';
 import * as Ethereum from '../../../utils/ethereum';
 import * as contractController from '../contract';
@@ -165,17 +166,12 @@ describe('contractController', () => {
 
     const unsubscribe1 = jest.fn();
     const unsubscribe2 = jest.fn();
-    const unsubscribe3 = jest.fn();
-    const unsubscribe4 = jest.fn();
 
     beforeEach(() => {
 
       contractController.contractSubscriptionList.push({
         eventEmitterEvents: {
           unsubscribe: unsubscribe1,
-        },
-        eventEmitterLogs: {
-          unsubscribe: unsubscribe2,
         },
         subscriptions: [{
           socketId: uuid,
@@ -186,10 +182,7 @@ describe('contractController', () => {
 
       contractController.contractSubscriptionList.push({
         eventEmitterEvents: {
-          unsubscribe: unsubscribe3,
-        },
-        eventEmitterLogs: {
-          unsubscribe: unsubscribe4,
+          unsubscribe: unsubscribe2,
         },
         subscriptions: [{
           socketId: uuid,
@@ -205,9 +198,7 @@ describe('contractController', () => {
 
       expect(contractController.contractSubscriptionList.length).toBe(1);
       expect(unsubscribe1).not.toHaveBeenCalled();
-      expect(unsubscribe2).not.toHaveBeenCalled();
-      expect(unsubscribe3).toHaveBeenCalled();
-      expect(unsubscribe4).toHaveBeenCalled();
+      expect(unsubscribe2).toHaveBeenCalled();
       expect(contractController.contractSubscriptionList[0].subscriptions.length).toBe(1);
 
     });
@@ -347,8 +338,6 @@ describe('contractController', () => {
 
     const unsubscribe1 = jest.fn();
     const unsubscribe2 = jest.fn();
-    const unsubscribe3 = jest.fn();
-    const unsubscribe4 = jest.fn();
 
     beforeEach(() => {
 
@@ -356,9 +345,6 @@ describe('contractController', () => {
         contractAddress: 'address',
         eventEmitterEvents: {
           unsubscribe: unsubscribe1,
-        },
-        eventEmitterLogs: {
-          unsubscribe: unsubscribe2,
         },
         subscriptions: [{
           socketId: uuid,
@@ -370,10 +356,7 @@ describe('contractController', () => {
       contractController.contractSubscriptionList.push({
         contractAddress: 'address2',
         eventEmitterEvents: {
-          unsubscribe: unsubscribe3,
-        },
-        eventEmitterLogs: {
-          unsubscribe: unsubscribe4,
+          unsubscribe: unsubscribe2,
         },
         subscriptions: [{
           socketId: uuid,
@@ -389,9 +372,7 @@ describe('contractController', () => {
 
       expect(contractController.contractSubscriptionList.length).toBe(1);
       expect(unsubscribe1).not.toHaveBeenCalled();
-      expect(unsubscribe2).not.toHaveBeenCalled();
-      expect(unsubscribe3).toHaveBeenCalled();
-      expect(unsubscribe4).toHaveBeenCalled();
+      expect(unsubscribe2).toHaveBeenCalled();
       expect(contractController.contractSubscriptionList[0].subscriptions.length).toBe(2);
 
     });
@@ -405,8 +386,6 @@ describe('contractController', () => {
       expect(contractController.contractSubscriptionList.length).toBe(2);
       expect(unsubscribe1).not.toHaveBeenCalled();
       expect(unsubscribe2).not.toHaveBeenCalled();
-      expect(unsubscribe3).not.toHaveBeenCalled();
-      expect(unsubscribe4).not.toHaveBeenCalled();
       expect(contractController.contractSubscriptionList[0].subscriptions.length).toBe(1);
 
     });
@@ -417,8 +396,6 @@ describe('contractController', () => {
 
     const subscribe1 = jest.fn();
     const subscribe2 = jest.fn();
-    const subscribe3 = jest.fn();
-    const subscribe4 = jest.fn();
     const on1 = jest.fn();
     const on2 = jest.fn();
     const allEventsMethod = jest.fn().mockReturnValueOnce({on: on1});
@@ -436,9 +413,6 @@ describe('contractController', () => {
         eventEmitterEvents: {
           subscribe: subscribe1,
         },
-        eventEmitterLogs: {
-          subscribe: subscribe2,
-        },
         subscriptions: [{
           socketId: uuid,
         }, {
@@ -454,10 +428,7 @@ describe('contractController', () => {
           },
         },
         eventEmitterEvents: {
-          subscribe: subscribe3,
-        },
-        eventEmitterLogs: {
-          subscribe: subscribe4,
+          subscribe: subscribe2,
         },
         subscriptions: [{
           socketId: uuid,
@@ -467,18 +438,16 @@ describe('contractController', () => {
 
     it('should call _restartSubscriptionsContracts correctly', async () => {
 
-      contractController.restartSubscriptionsContracts();
+      contractController.restartSubscriptionsContracts({});
 
       expect(subscribe1).not.toHaveBeenCalled();
-      expect(subscribe3).not.toHaveBeenCalled();
-      expect(subscribe2).toHaveBeenCalled();
-      expect(subscribe4).toHaveBeenCalled();
+      expect(subscribe2).not.toHaveBeenCalled();
       expect(allEventsMethod).toHaveBeenCalled();
       expect(allEventsMethod2).toHaveBeenCalled();
       expect(on1).toHaveBeenCalled();
       expect(on2).toHaveBeenCalled();
-
     });
+
   });
 
   describe('_processEvent', () => {
@@ -496,7 +465,7 @@ describe('contractController', () => {
       transactionHash: 'hash',
       address: 'scAddress',
       blockNumber: 0,
-      event: undefined,
+      event: 'Transfer',
       id: 'log_5daf9707',
       logIndex: 0,
       raw: {
@@ -509,13 +478,28 @@ describe('contractController', () => {
       type: 'mined',
     };
     const blockHeader = {
+      timestamp: 100,
       transactions: [
         {
           hash: 'hash',
-          gas: 'gas',
-          gasPrice: 'hash',
+          gas: 100,
+          gasPrice: '20',
         },
       ],
+    };
+    const generatedBody = {
+      blockNumber: eventBody.blockNumber,
+      blockHash: eventBody.blockHash,
+      transactionId: eventBody.transactionHash,
+      smartContractAddress: eventBody.address,
+      eventName: eventBody.event,
+      returnValues: eventBody.returnValues,
+      fee: {
+        amount: '2000',
+        decimals: 18,
+        currency: CURRENCY.Ethereum,
+      },
+      timestamp: 100,
     };
     let _getBlock: any;
 
@@ -533,7 +517,13 @@ describe('contractController', () => {
 
       await contractController._processEvent(sub, web3I, eventBody);
 
-      expect(notify).toHaveBeenCalledTimes(2);
+      const response = {
+        kind: CONSUMER_EVENT_KINDS.SmartContractEvent,
+        body: generatedBody,
+        raw: eventBody,
+        matchedAddress: eventBody.address,
+      };
+      expect(notify).toHaveBeenNthCalledWith(1, response);
     });
 
   });