+++ src/models/models.ts
@@ -45,6 +45,7 @@ export interface ISocketEvent {
   kind: ISocketEventKind;
   body: ISocketEventBody;
   matchedAddress?: dltAddress;
+  raw?: ISocketEventBody;
 }
 
 export enum CONSUMER_EVENT_KINDS {
@@ -52,4 +53,37 @@ export enum CONSUMER_EVENT_KINDS {
   Transaction = 'transaction',
   SmartContractTransaction = 'contract-transaction',
   SmartContractEvent = 'contract-event',
+}
+
+export interface IHancockSocketTransactionBody extends ISocketEventBody {
+  blockHash: string;
+  blockNumber: number;
+  transactionId: string;
+  from: string;
+  to: string;
+  value: IHancockSocketCurrency;
+  data: string;
+  fee: IHancockSocketCurrency;
+  timestamp: number;
+}
+
+export interface IHancockSocketCurrency {
+  amount: string;
+  decimals: number;
+  currency: CURRENCY;
+}
+
+export enum CURRENCY {
+  Ethereum = 'Ethereum',
+}
+
+export interface IHancockSocketContractEventBody extends ISocketEventBody {
+  blockNumber: number;
+  blockHash: string;
+  transactionId: string;
+  smartContractAddress: string;
+  eventName: string;
+  returnValues: string[];
+  fee: IHancockSocketCurrency;
+  timestamp: number;
 }